+++ CHANGES.md
@@ -5,11 +5,13 @@ E-Maj - Change log
 ###Enhancements:###
   * When launching an E-Maj rollback, allow to set a comment about the
     operation.
+  * Add a emaj_comment_rollback() function to set/change/reset a comment on
+    an E-Maj rollback.
+  * In the E-Maj rollback execution report, add a Notice message indicating
+    the rollback id.
   * Record the E-Maj rollback start, initialization end and locking end
     timestamps into the emaj_rlbk table and let the emaj_rollback_activity()
     function report both rollback planning and locking phases duration.
-  * In the E-Maj rollback execution report, add a Notice message indicating
-    the rollback id.
   * Minor code changes.
 
 ###Bug fixes:###